move constant.py file
