+++ swap_comparator/constant.py
